+++ src/Nethermind/Nethermind.State.Test/StateProviderTests.cs
@@ -88,18 +88,6 @@ namespace Nethermind.Store.Test
         }
 
         [Test]
-        public void Can_collect_stats()
-        {
-            WorldState provider = new(new TrieStore(new MemDb(), Logger), _codeDb, Logger);
-            provider.CreateAccount(TestItem.AddressA, 1.Ether());
-            provider.Commit(MuirGlacier.Instance);
-            provider.CommitTree(0);
-
-            var stats = provider.CollectStats(_codeDb, Logger);
-            stats.AccountCount.Should().Be(1);
-        }
-
-        [Test]
         public void Can_accepts_visitors()
         {
             WorldState provider = new(new TrieStore(new MemDb(), Logger), Substitute.For<IDb>(), Logger);
